[OGUI-368] Split mocha tests in multiple files
